add chatbot as componnet
