+++ package.json
@@ -36,6 +36,7 @@
     "@vercel/analytics": "^1.3.1",
     "@vercel/kv": "^3.0.0",
     "@vercel/speed-insights": "^1.0.12",
+    "build": "^0.1.4",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
